# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",   
    "version": "1.0.2",
    "title": "testdb",
    "termsOfService": "",
    "contact": {
      "email": "shubhamjulaha.rexweb@gmail.com"
    },
    "license": {
      "name": "",
      "url": ""
    }
  },
  "host": "192.168.43.50:3000",
  "basePath": "",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/mom": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "MOM Module"
        ],
        "summary": "This API will return all Minutes Of Meeting Record",
        "description": "List of all MOM fields",
        "operationId": "getMOM fields",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "MOM Module"
        ],
        "summary": " This API is for adding new MOM entry in tblmomfields ",
        "description": "",
        "operationId": "addMOMfields",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": [
              "email"
            ],
            "schema": {
              "$ref": "#/definitions/MOM"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Please fill all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
...
    "/api/hrcandidacy": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "HR Candidacy Module"
        ],
        "summary": "This API will display list of all candidates appeared in an interview. ",
        "description": "This module is for maintaining relationship with clients",
        "operationId": "getcandidacyfields",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "HR Candidacy Module"
        ],
        "summary": "This API will add a new candidate record in candidancy module  ",
        "description": "Add a new field into db",
        "operationId": "addcandidacy",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/HrCandidacy"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Please fill all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "MOM": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "reasonOfMeeting": {
          "type": "string"
        },
        "dateOfMom": {
          "type": "string"
        },
        "durationMeeting": {
          "type": "string"
        },
        "initiated": {
          "type": "string"
        },
        "participants": {
          "type": "string"
        },
        "clientName": {
          "type": "string"
        },
        "medium": {
          "type": "string"
        },
        "communication": {
          "type": "string"
        },
        "clientInput": {
          "type": "string"
        },
        "development": {
          "type": "string"
        },
        "core": {
          "type": "string"
        },
        "designing": {
          "type": "string"
        },
        "hr": {
          "type": "string"
        },
        "approved": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
...
    "HrCandidacy": {
      "type": "object",
      "properties": {
        "candidateName": {
          "type": "string"
        },
        "experience": {
          "type": "string"
        },
        "currentSalary": {
          "type": "integer"
        },
        "expectedSalary": {
          "type": "integer"
        },
        "noticePeriod": {
          "type": "string"
        },
        "interviewDate": {
          "type": "string"
        },
        "interviewTime": {
          "type": "string"
        },
        "interviewMode": {
          "type": "string"
        },
        "reasonForChange": {
          "type": "string"
        },
        "outcome": {
          "type": "string"
        },
        "isActive": {
          "type": "integer"
        }
      }
    },
...
  },
...
}